dash_app/requirements.txt
dash
dash-bootstrap-components
plotly
torch
torchvision
timm
opencv-python
matplotlib
scikit-learn
Pillow
streamlit
pyttsx3
gtts
numpy
pandas
seaborn
dash
dash-bootstrap-components
plotly
torch
torchvision
timm
opencv-python
ultralytics
numpy
pandas
scikit-learn
matplotlib
seaborn
pyttsx3
gTTS
reportlab
pdfkit
albumentations